🚚 Move generic function to utils
